add panorama viewer to lightbox
