+++ libraries/mapping/lightbox/src/index.ts
@@ -4,6 +4,9 @@ export type { CarmaLightBoxProps } from "./lib/CarmaLightBox";
 export { default as PhotoLightBox } from "./lib/PhotoLightBox";
 export type { PhotoLightBoxProps } from "./lib/PhotoLightBox";
 
+export { default as MediaLightBox } from "./lib/MediaLightBox";
+export type { MediaLightBoxProps } from "./lib/MediaLightBox";
+
 export { default as InfoBoxFotoPreview } from "./lib/InfoBoxFotoPreview";
 export type { InfoBoxFotoPreviewProps } from "./lib/InfoBoxFotoPreview";
 
@@ -18,6 +21,9 @@ export type {
   LightBoxDispatchValue,
   LightBoxSetAllPayload,
   LightBoxContextProviderProps,
+  LightBoxSlide,
+  LightBoxImageSlide,
+  LightBoxCustomSlide,
 } from "./lib/LightBoxContextProvider";
 
 export {
@@ -25,5 +31,6 @@ export {
   getLinkOrText,
   fotoKraemerUrlManipulation,
   fotoKraemerCaptionFactory,
+  defaultLightBoxCaptionFactory,
 } from "./lib/lightboxHelpers";
 export type { TriggerLightBoxForFeatureArgs } from "./lib/lightboxHelpers";
